+++ Core/Src/I2C_IT.cpp
@@ -8,7 +8,7 @@
 #include "I2C_IT.h"
 //Select I2C instance no from {1,2,3} (STM32NucleoL476 has 3 instances I2C1, I2C2, I2C3) and mode from {1=normal(100kbps), 2=fastmode(400kbps), 3=fastmodeplus(1Mbps)}
 
-I2C_IT::I2C_IT(uint8_t tempaddress, I2C_HandleTypeDef hi2c, uint32_t delay )
+I2C_IT::I2C_IT(uint16_t tempaddress, I2C_HandleTypeDef hi2c, uint32_t delay )
 {
 address=tempaddress;
 hi2c1=hi2c;
@@ -16,7 +16,7 @@ Wait_Delay=delay;
 
 }
 
-I2C_STATUS I2C_IT::Transmit(uint8_t (&bits)[], int no_of_bytes)
+I2C_STATUS I2C_IT::Transmit(uint8_t (&bits)[], uint16_t no_of_bytes)
 {
 	HAL_StatusTypeDef temp;
 	temp= HAL_I2C_Master_Transmit_IT(&hi2c1 , address, bits, no_of_bytes);
@@ -68,7 +68,7 @@ void I2C::Transmit_DMA(uint8_t *bits, int no_of_bytes)
 
 
 
-I2C_STATUS I2C_IT::Receive_2_Buffer(uint8_t (&I2C_Buffer)[],int no_of_bytes)
+I2C_STATUS I2C_IT::Receive_2_Buffer(uint8_t (&I2C_Buffer)[],uint16_t no_of_bytes)
 {
 	HAL_StatusTypeDef temp;
 		temp= HAL_I2C_Master_Receive_IT(&hi2c1, address, I2C_Buffer, no_of_bytes);
